is_dense trait moved to fwd.h
