+++ src/TiledArray/fwd.h
@@ -147,6 +147,23 @@ class SparseShape;
 template <typename, typename>
 class DistArray;
 
+/// Type trait to detect dense shape types
+template <typename S>
+struct is_dense : public std::false_type {};
+
+template <>
+struct is_dense<DenseShape> : public std::true_type {};
+
+template <>
+struct is_dense<DensePolicy> : public std::true_type {};
+
+template <typename Tile, typename Policy>
+struct is_dense<DistArray<Tile, Policy>>
+    : public is_dense<typename DistArray<Tile, Policy>::shape_type> {};
+
+template <typename T>
+constexpr const bool is_dense_v = is_dense<T>::value;
+
 // Dense Array Typedefs
 template <typename T>
 using TArray = DistArray<Tensor<T>, DensePolicy>;
